Rename class Input as FileReader and implement changes
