+++ src/main/java/com/codecool/battleofcards/card/Deck.java
@@ -1,6 +1,6 @@
 package com.codecool.battleofcards.card;
 
-import com.codecool.battleofcards.input.*;
+import com.codecool.battleofcards.reader.*;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -11,7 +11,7 @@ import java.util.Scanner;
 
 public class Deck implements Shuffleable {
     private List<Card> cards;
-    Input input = new Input();
+    FileReader fileReader = new FileReader();
 
     public Deck(String filename){
         this.cards = new ArrayList<>();
@@ -19,7 +19,7 @@ public class Deck implements Shuffleable {
     }
 
     private void createCards(String filename){
-        String[] statisticsList = input.readerFromFile(filename).split("\n");
+        String[] statisticsList = fileReader.readerFromFile(filename).split("\n");
         
         for (String line : statisticsList){
             String[] cardStatistic = line.split(",");
